test(perf): Separates unit and perf tests
Currently there are very few performance tests but these changes separate them from the unit tests
so that they can grow independently.

To run unit tests you still `npm test` and for performance tests you now `npm run test:perf`.

Closes #45
