+++ spec/unit/matlab/conv2.spec.js
@@ -359,72 +359,3 @@ test('convolutions with box kernels should default to "full"', (t) => {
 	t.deepEqual(defaultOut, fullOut);
 	t.end();
 });
-
-test('box kernels should perform better than non box kernels', (t) => {
-	const A = {
-		data: [
-			10, 20, 30, 10, 20, 30, 10, 20, 30,
-			40, 50, 60, 40, 50, 60, 40, 50, 60,
-			70, 80, 90, 70, 80, 90, 70, 80, 90,
-			10, 20, 30, 10, 20, 30, 10, 20, 30,
-			40, 50, 60, 40, 50, 60, 40, 50, 60,
-			70, 80, 90, 70, 80, 90, 70, 80, 90,
-			10, 20, 30, 10, 20, 30, 10, 20, 30,
-			40, 50, 60, 40, 50, 60, 40, 50, 60,
-			70, 80, 90, 70, 80, 90, 70, 80, 90
-		],
-		width: 9,
-		height: 9
-	};
-	const box = {
-		data: [
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 9
-		],
-		width: 9,
-		height: 9
-	};
-	const nonBox = {
-		data: [
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9,
-			1 / 9, 1 / 9, 1 / 9, 1 / 9, 1 / 8, 1 / 9, 1 / 9, 1 / 9, 1 / 9
-		],
-		width: 9,
-		height: 9
-	};
-	const repetitions = 1000;
-
-	let start = new Date().getTime();
-
-	for (let i = 0; i < repetitions; i++) {
-		conv2(A, box, 'full');
-	}
-
-	const boxTime = new Date().getTime() - start;
-
-	start = new Date().getTime();
-
-	for (let i = 0; i < repetitions; i++) {
-		conv2(A, nonBox, 'full');
-	}
-
-	const nonBoxTime = new Date().getTime() - start;
-
-	t.equal(boxTime < nonBoxTime, true,
-		`box kernel (${boxTime}) should be faster than non box kernel (${nonBoxTime})`);
-	t.end();
-});
